pyproject.toml
[tool.poetry]
name = "lms-drf"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
black = "^24.8.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.6.1"
django-cors-headers = "^4.4.0"
stripe = "^10.9.0"
currencyconverter = "^0.17.30"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
